Implementaciones de Login y verificacion de tipo de empresa durante el login para distinta pagina segun tipo de empresa del usuario.

+++ RSWork-Backend/Empresa.cs
@@ -9,7 +9,7 @@ namespace BE
     public abstract class Empresa
     {
 
-        public int CUIT { get; set; }
+        public Int64 CUIT { get; set; }
 
         public string Direccion { get; set; }
